requirements.txt
python-memcached
kinetic
